+++ src/cc-context/index.js
@@ -3,7 +3,7 @@ import computed from './computed-map';
 import watch from './watch-map';
 import runtimeVar from './runtime-var';
 import { waKey_uKeyMap_, waKey_staticUKeyMap_ } from './wakey-ukey-map';
-import { MODULE_GLOBAL, MODULE_CC, MODULE_DEFAULT, MODULE_VOID, CATE_MODULE, CC_DISPATCHER } from '../support/constant';
+import { MODULE_GLOBAL, MODULE_CC, MODULE_DEFAULT, MODULE_VOID, CATE_MODULE } from '../support/constant';
 import * as util from '../support/util';
 import pickDepFns from '../core/base/pick-dep-fns';
 import findDepFnsToExecute from '../core/base/find-dep-fns-to-execute';
@@ -12,7 +12,7 @@ import extractStateByKeys from '../core/state/extract-state-by-keys';
 const { _computedValue } = computed;
 const { okeys, extractChangedState } = util;
 const refs = {};
-const getDispatcher = () => refs[CC_DISPATCHER];
+const getDispatcher = () => ccContext.permanentDispatcher;
 
 const setStateByModule = (module, committedState, { ref = null, callInfo = {}, noSave = false } = {}) => {
   const moduleState = getState(module);
@@ -27,18 +27,16 @@ const setStateByModule = (module, committedState, { ref = null, callInfo = {}, n
   const callerRef = ref || getDispatcher();
   const refModule = callerRef.module;
   const newState = Object.assign({}, moduleState, committedState);
-
   const deltaCommittedState = Object.assign({}, committedState);
-  let stateForComputeFn = deltaCommittedState;
 
-  findDepFnsToExecute(
+  const { hasDelta: hasDeltaInCu } = findDepFnsToExecute(
     callerRef, module, refModule, moduleState, curDepComputedFns,
-    stateForComputeFn, newState, deltaCommittedState, callInfo, false,
+    deltaCommittedState, newState, deltaCommittedState, callInfo, false,
     'computed', CATE_MODULE, moduleComputedValue,
   );
-  findDepFnsToExecute(
+  const { hasDelta: hasDeltaInWa } = findDepFnsToExecute(
     callerRef, module, refModule, moduleState, curDepWatchFns,
-    stateForComputeFn, newState, deltaCommittedState, callInfo, false,
+    deltaCommittedState, newState, deltaCommittedState, callInfo, false,
     'watch', CATE_MODULE, moduleComputedValue,
   );
 
@@ -46,7 +44,10 @@ const setStateByModule = (module, committedState, { ref = null, callInfo = {}, n
     saveSharedState(module, deltaCommittedState);
   }
 
-  return deltaCommittedState;
+  return {
+    hasDelta: hasDeltaInCu || hasDeltaInWa,
+    deltaCommittedState,
+  };
 }
 
 const saveSharedState = (module, toSave, needExtract = false) => {
@@ -58,6 +59,8 @@ const saveSharedState = (module, toSave, needExtract = false) => {
 
   const moduleState = getState(module);
   const prevModuleState = getPrevState(module);
+  incModuleVer(module);
+
   return extractChangedState(moduleState, target, {
     prevStateContainer: prevModuleState,
     incStateVer: key => incStateVer(module, key),
@@ -70,6 +73,19 @@ const getState = (module) => {
 
 const getPrevState = (module) => {
   return _prevState[module];
+}
+
+const getModuleVer = function (module) {
+  if (!module) return _moduleVer;
+  return _moduleVer[module];
+}
+
+const incModuleVer = function (module) {
+  try {
+    _moduleVer[module]++;
+  } catch (err) {
+    _moduleVer[module] = 1;
+  }
 }
 
 const getStateVer = function (module) {
@@ -94,6 +110,9 @@ const _prevState = getRootState();
 // 1 effect里的函数再次出发当前实例渲染，渲染完后检查prevModuleState curModuleState, 对应的key值还是不一样，又再次出发effect，造成死循环
 // 2 确保引用型值是基于原有引用修改某个属性的值时，也能触发effect
 const _stateVer = {};
+// 优化before-render里无意义的merge mstate导致冗余的set（太多的set会导致 Maximum call stack size exceeded）
+// https://codesandbox.io/s/happy-bird-rc1t7?file=/src/App.js concent below 2.4.18会触发
+const _moduleVer = {};
 
 const ccContext = {
   getDispatcher,
@@ -120,26 +139,28 @@ const ccContext = {
   moduleName_isConfigured_: {
   },
   /**
-    ccClassContext:{
-      module,
-      ccClassKey,
-      // renderKey机制影响的类范围，默认只影响调用者所属的类，如果有别的类观察了同一个模块的某个key，这个类的实例是否触发渲染不受renderKey影响
-      // 为 * 表示影响所有的类，即其他类实例都受renderKey机制影响。
-      renderKeyClasses, 
-      originalWatchedKeys,
-      watchedKeys,
-      ccKeys: [],
-      connectedState: {},
-      connectedModuleKeyMapping: null,
-      connectedModule:{},//记录当前cc类连接到了其他哪些模块
-    }
+   * ccClassContext:{
+   *   module,
+   *   ccClassKey,
+   *   // renderKey机制影响的类范围，默认只影响调用者所属的类，如果有别的类观察了同一个模块的某个key，这个类的实例是否触发渲染不受renderKey影响
+   *   // 为 * 表示影响所有的类，即其他类实例都受renderKey机制影响。
+   *   renderKeyClasses, 
+   *   originalWatchedKeys,
+   *   watchedKeys,
+   *   ccKeys: [],
+   *   connectedState: {},
+   *   connectedModuleKeyMapping: null,
+   *   connectedModule:{},//记录当前cc类连接到了其他哪些模块
+   * }
   */
   ccClassKey_ccClassContext_: {
   },
-  // globalStateKeys is maintained by cc automatically,
-  // when user call cc.setGlobalState, or ccInstance.setGlobalState,
-  // commit state will be checked strictly by cc with globalStateKeys,
-  // all the keys of commit state must been included in globalStateKeys
+  /**
+   * globalStateKeys is maintained by cc automatically,
+   * when user call cc.setGlobalState, or ccInstance.setGlobalState,
+   * committedState will be checked strictly by cc with globalStateKeys,
+   * committedState keys must been included in globalStateKeys
+   */
   globalStateKeys: [
   ],
   //store里的setState行为会自动触发模块级别的computed、watch函数
@@ -165,6 +186,7 @@ const ccContext = {
       else return _prevState;
     },
     getStateVer,
+    getModuleVer,
     setState: function (module, partialSharedState, options) {
       return setStateByModule(module, partialSharedState, options);
     },
@@ -203,14 +225,17 @@ const ccContext = {
     _init: {}
   },
   ccUKey_ref_: refs,
-  //  key:eventName,  value: Array<{ccKey, identity,  handlerKey}>
+  /**
+   * key:eventName,  value: Array<{ccKey, identity,  handlerKey}>
+   */
   event_handlers_: {},
   ccUKey_handlerKeys_: {},
-  // to avoid memory leak, the handlerItem of event_handlers_ just store handlerKey, 
-  // it is a ref that towards ccUniqueKeyEvent_handler_'s key
-  // when component unmounted, its handler will been removed
+  /**
+   * to avoid memory leak, the handlerItem of event_handlers_ just store handlerKey, 
+   * it is a ref that towards ccUniqueKeyEvent_handler_'s key
+   * when component unmounted, its handler will been removed
+   */
   handlerKey_handler_: {},
-  // { 'foo/f1': {ukey1: 1, ukey2:1 } }
   waKey_uKeyMap_,
   waKey_staticUKeyMap_,
   refs,
@@ -218,37 +243,43 @@ const ccContext = {
     packageLoadTime: Date.now(),
     firstStartupTime: '',
     latestStartupTime: '',
-    version: '2.4.15',
+    version: '2.6.1',
     author: 'fantasticsoul',
     emails: ['624313307@qq.com', 'zhongzhengkai@gmail.com'],
     tag: 'yuna',
   },
 
-  // fragment association
-  fragmentNameCount: 0,
   featureStr_classKey_: {},
   userClassKey_featureStr_: {},
   errorHandler: null,
   middlewares: [],
   plugins: [],
   pluginNameMap: {},
+  permanentDispatcher: null,
+  localStorage: null,
+  recoverRefState: () => { },
+}
+
+ccContext.recoverRefState = function () {
+  const localStorage = ccContext.localStorage;
+  if (!localStorage) return;
+
+  const lsLen = localStorage.length;
+  const _refStoreState = ccContext.refStore._state;
+  for (let i = 0; i < lsLen; i++) {
+    const lsKey = localStorage.key(i);
+    if (lsKey.startsWith('CCSS_')) {
+      try {
+        _refStoreState[lsKey.substr(5)] = JSON.parse(localStorage.getItem(lsKey));
+      } catch (err) {
+        console.error(err);
+      }
+    }
+  }
 }
 
 export function getCcContext() {
   return ccContext;
-}
-
-const lsLen = localStorage.length;
-const _refStoreState = ccContext.refStore._state;
-for (let i = 0; i < lsLen; i++) {
-  const lsKey = localStorage.key(i);
-  if (lsKey.startsWith('CCSS_')) {
-    try {
-      _refStoreState[lsKey.substr(5)] = JSON.parse(localStorage.getItem(lsKey));
-    } catch (err) {
-      console.error(err);
-    }
-  }
 }
 
 export default ccContext;