+++ src/core/base/map-registration-info.js
@@ -87,12 +87,12 @@ export default function (
   checker.checkModuleName(module, false, `module[${module}] is not configured in store`);
   checker.checkStoredKeys(moduleName_stateKeys_[module], inputStoredKeys);
 
-  let _connect = connect;
+  let _connect = connect || {};// codesandbox lost default value
   const isArr = Array.isArray(connect);
   if (isArr || typeof connect === 'string') {
     _connect = {};
     const connectedModules = isArr ? connect : connect.split(',');
-    connectedModules.forEach(m => _connect[m] = '-');//标识自动收集观察依赖
+    connectedModules.forEach(m => { _connect[m] = '-' });//标识自动收集观察依赖
   }
 
   // 不指定global模块的话，默认自动收集global观察依赖，方便用户直接使用ctx.globalState时，就触发自动收集