+++ src/core/base/sync.js
@@ -20,10 +20,11 @@ export default function (spec, ref, e) {
 
   if (e && e.stopPropagation) e.stopPropagation();
 
-  const { ccsync, ccint, ccdelay, ccrkey } = dataset;
+  const { ccint, ccdelay, ccrkey } = dataset;
+  let ccsync = dataset.ccsync;
 
   if (ccsync.startsWith('/')) {
-    dataset.ccsync = `${refModule}${ccsync}`;//附加上默认模块值
+    ccsync = `${refModule}${ccsync}`;//附加上默认模块值
   }
 
   if (ccsync.includes('/')) {// syncModuleState 同步模块的state状态
@@ -31,15 +32,17 @@ export default function (spec, ref, e) {
     checker.checkModuleName(targetModule, false);
 
     const { ccKey, ccUniqueKey } = refCtx;
+    const options = { calledBy: SYNC, ccKey, ccUniqueKey, module: targetModule, renderKey: ccrkey, delay: ccdelay };
+
     if (noAutoExtract) {
-      if (extraState) changeRefState(extraState, { calledBy: SYNC, ccKey, ccUniqueKey, module: targetModule, renderKey: ccrkey, delay: ccdelay }, ref);
+      if (extraState) changeRefState(extraState, options, ref);
       return;
     }
 
     const fullState = targetModule !== refModule ? getState(targetModule) : ref.state;
 
     const { state } = extractStateByCcsync(ccsync, value, ccint, fullState, mockE.isToggleBool);
-    changeRefState(state, { calledBy: SYNC, ccKey, ccUniqueKey: ccUniqueKey, module: targetModule, renderKey: ccrkey, delay: ccdelay }, ref);
+    changeRefState(state, options, ref);
   } else {//调用自己的setState句柄触发更新，key可能属于local的，也可能属于module的
     if (noAutoExtract) {
       if (extraState) ref.setState(extraState, null, ccrkey, ccdelay);